+++ packages/custom-picker-button/src/mobile/index.ts
@@ -1,0 +1,1 @@
+export { CustomPickerButtonMobile, CustomPickerButtonMobileProps } from '../Component.mobile';