+++ packages/input-autocomplete/src/desktop.ts
@@ -1,1 +1,0 @@
-export { InputAutocompleteDesktop, InputAutocompleteDesktopProps } from './Component.desktop';