+++ packages/input-autocomplete/src/mobile.ts
@@ -1,5 +1,0 @@
-export { InputAutocompleteMobile, InputAutocompleteMobileProps } from './Component.mobile';
-export {
-    InputAutocompleteModalMobile,
-    InputAutocompleteModalMobileProps,
-} from './Component.modal.mobile';