+++ packages/modal/src/mobile/index.ts
@@ -1,0 +1,2 @@
+export * from '../Component.mobile';
+export { ModalMobileProps } from '../typings';