+++ packages/number-input/src/desktop.ts
@@ -1,1 +1,0 @@
-export { NumberInputDesktop, NumberInputDesktopProps } from './Component.desktop';