+++ packages/picker-button/src/mobile.ts
@@ -1,1 +1,0 @@
-export { PickerButtonMobile, PickerButtonMobileProps } from './Component.mobile';