+++ packages/picker-button/src/mobile/index.ts
@@ -1,0 +1,1 @@
+export { PickerButtonMobile, PickerButtonMobileProps } from '../Component.mobile';