+++ packages/select/src/desktop.ts
@@ -1,1 +1,0 @@
-export { SelectDesktop, SelectDesktopProps } from './Component.desktop';