+++ packages/select/src/shared.ts
@@ -1,5 +1,0 @@
-export * from './components';
-export * from './presets';
-export * from './typings';
-export * from './utils';
-export type { AdditionalMobileProps } from './Component.mobile';