+++ packages/tabs/src/collapsible/index.ts
@@ -1,0 +1,2 @@
+export { TabsCollapsible, TabsCollapsibleProps } from '../components/tabs/Component.collapsible';
+export * from '../components/tab';