+++ packages/tabs/src/desktop.ts
@@ -1,4 +1,0 @@
-export * from './components/primary-tablist/Component.desktop';
-export * from './components/secondary-tablist/Component.desktop';
-export * from './components/tab';
-export * from './components/tabs/Component.desktop';