+++ packages/tabs/src/mobile.ts
@@ -1,4 +1,0 @@
-export * from './components/primary-tablist/Component.mobile';
-export * from './components/secondary-tablist/Component.mobile';
-export * from './components/tab';
-export * from './components/tabs/Component.mobile';